Refactor: combine DELETE and GET query params
